move middleware to its own folder
