+++ src/store/middleware/index.js
@@ -1,0 +1,3 @@
+import promiseCatchAll from './promise-catch-all';
+
+export default { promiseCatchAll };
